APP-002 T78  & APP-003 T34-36

+++ FrontEnd/finalproject_edunamica/src/Services/Courses/DeleteCourses.jsx
@@ -1,9 +1,19 @@
 async function DeleteCourse(courseId) {
+
+    const token = localStorage.getItem('access_token');
+    if (!token) {
+        console.error("No token found");
+        return;
+    }
+
+    const validationToken = `Bearer ${token}`;
+
     try {
         const response = await fetch(`http://localhost:8000/api/courses/${courseId}/`, {
             method: 'DELETE',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization': validationToken,
             }
         });
